chore: moved article model to lib
