+++ article-service/main.go
@@ -5,16 +5,14 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"time"
 
-	"github.com/google/uuid"
-	"github.com/lib/pq"
 	"github.com/rs/zerolog/log"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
+	"github.com/leonsteinhaeuser/example-app/lib"
 )
 
 var (
@@ -28,27 +26,6 @@ var (
 
 	gormDB *gorm.DB
 )
-
-type Article struct {
-	// identifier and state fields
-	ID        uuid.UUID `gorm:"type:uuid;default:uuid_generate_v4();primaryKey"`
-	CreatedAt time.Time
-	UpdatedAt time.Time
-	DeletedAt gorm.DeletedAt `gorm:"index"`
-
-	// model fields
-
-	// Title reprGormDBDataTypeesents the title of the article
-	Title string `json:"title,omitempty"`
-	// Description represents a short description what the article is about
-	Description string `json:"description,omitempty"`
-	// Author represents the author of the article
-	AuthorID uuid.UUID `json:"author_id,omitempty" gorm:"type:uuid"`
-	// Text is the actual article text
-	Text string `json:"text,omitempty"`
-
-	KeywordIDs pq.StringArray `json:"keyword_ids,omitempty" gorm:"type:uuid[]"`
-}
 
 func init() {
 	log.Info().Msg("initializing database")
@@ -78,7 +55,7 @@ func init() {
 	}
 
 	log.Info().Msg("migrating database tables...")
-	err = gormDB.AutoMigrate(Article{})
+	err = gormDB.AutoMigrate(lib.Article{})
 	if err != nil {
 		log.Fatal().Err(err).Msg("failed to migrate database table")
 	}
@@ -134,7 +111,7 @@ func getByID(w http.ResponseWriter, r *http.Request) {
 	id := getURLID(r)
 
 	log.Debug().Msg("requesting article by ID from database")
-	article := &Article{}
+	article := &lib.Article{}
 	err := gormDB.Model(article).First(article, "id = ?", id).Error
 	if err != nil {
 		log.Error().Err(err).Msg("failed to article by id")
@@ -155,7 +132,7 @@ func getByID(w http.ResponseWriter, r *http.Request) {
 }
 
 func list(w http.ResponseWriter, r *http.Request) {
-	articles := &[]Article{}
+	articles := &[]lib.Article{}
 
 	log.Debug().Msg("querying database for all articles")
 	err := gormDB.Model(articles).Find(articles).Error
@@ -181,7 +158,7 @@ func updateByID(w http.ResponseWriter, r *http.Request) {
 	id := getURLID(r)
 
 	log.Debug().Msg("parsing request from json to article")
-	article := Article{}
+	article := lib.Article{}
 	err := json.NewDecoder(r.Body).Decode(&article)
 	if err != nil {
 		log.Error().Err(err).Msg("failed to decode request body")
@@ -210,7 +187,7 @@ func deleteByID(w http.ResponseWriter, r *http.Request) {
 	id := getURLID(r)
 
 	log.Debug().Msg("unparsing request body to article")
-	article := &Article{}
+	article := &lib.Article{}
 	err := json.NewDecoder(r.Body).Decode(article)
 	if err != nil {
 		log.Error().Err(err).Msg("failed to decode request body")
@@ -237,7 +214,7 @@ func deleteByID(w http.ResponseWriter, r *http.Request) {
 
 func create(w http.ResponseWriter, r *http.Request) {
 	log.Debug().Msg("unparsing request body to article")
-	article := &Article{}
+	article := &lib.Article{}
 	err := json.NewDecoder(r.Body).Decode(article)
 	if err != nil {
 		log.Error().Err(err).Msg("failed to decode request body")
